Rename some audio
Add audio feedback for hand absorb error

Plug socket interactor when no stacks left

+++ Assets/Code/Unity/HandAbsorber/HandAbsorber.cs
@@ -1,5 +1,4 @@
-using System.Collections;
-
+using SaloonSlingers.Core;
 using SaloonSlingers.Unity.Actor;
 
 using UnityEngine;
@@ -8,15 +7,35 @@ namespace SaloonSlingers.Unity
 {
     public class HandAbsorber : MonoBehaviour
     {
-        [SerializeField]
-        private float absorbTime = 0.5f;
+        public Points Stacks { get; private set; }
+        public bool CanAbsorb { get => Stacks > 0; }
 
-        public IEnumerator Absorb(HitPoints hitPoints, HandProjectile projectile)
+        [SerializeField]
+        private uint startingStacks = 3;
+
+        private void Awake()
         {
+            Stacks = new Points(startingStacks);
+        }
+
+        public void Absorb(HitPoints hitPoints, HandProjectile projectile)
+        {
+            if (!CanAbsorb) return;
+
+            hitPoints.Points.Decreased += CheckStackRegained;
+            Stacks.Decrement();
+
             projectile.Pause();
-            yield return new WaitForSeconds(absorbTime);
-            hitPoints.Points.Reset(projectile.HandEvaluation.Score);
+            hitPoints.Points.Reset(hitPoints + projectile.HandEvaluation.Score);
             projectile.Kill();
+        }
+
+        private void CheckStackRegained(Points sender, ValueChangeEvent<uint> e)
+        {
+            if (e.After != 0) return;
+
+            Stacks.Increment();
+            sender.Decreased -= CheckStackRegained;
         }
     }
 }
